wip: merge editingLayerInner to editingLayer
